rename ThemeCard to MessageTemplateCard
